M-1 + T-25 Mechanical Changes (#3638)
* M-1 + T-25 Balance Pass

Some new visuals for new projectile types-- the Kinetic Bolt on the M1 and the Flare Shot on the T25. Some stat adjustments as well.

* Finalizing ammo changes

I think, 8 shots is actually a bit much and bumped it down to 7. For the magnum as well, I think with the self-recharge being back at 31 seconds with only a 1 second delay I think keeping it at 6 should be fine. Other ammo types remain the same.