SecLink Utility Expansion (#3540)
* SecLink Utility Expansion

Adding some big things to the Utility section of the SecLink to flesh out options, as well as removing the Portable Criminal Records from the list of options.

* Handheld in the HoS Locker

puts the Handheld in the HoS Locker

* Actually I prefer this

Making the handheld an on-spawn item rather than a locker item. HoS locker is a major loot box as it is, and it doesn't need to be made even more dramatic to be broken into

* This too

Sec Shield breaks in 5 shots (35 damage)

* description changes

changing the shield and broken shield's descriptions